Removed old Queries Spec
