+++ bounded-test/src/main/scala/io/cafienne/bounded/test/TestableProjection.scala
@@ -45,7 +45,9 @@ class TestableProjection private (system: ActorSystem, timeout: Timeout) {
 
   val eventStreamListener = TestProbe()
 
-  if (!system.settings.config.hasPath("io.bounded.eventmaterializers.publish") || !system.settings.config.getBoolean("io.bounded.eventmaterializers.publish")) {
+  if (!system.settings.config.hasPath("io.bounded.eventmaterializers.publish") || !system.settings.config.getBoolean(
+        "io.bounded.eventmaterializers.publish"
+      )) {
     system.log.error("Config property io.bounded.eventmaterializers.publish must be enabled")
   }
